release v1.0.101 (versionCode 134): CHECK_REPORT 66 项全量复查 + 编译链修复
- 修复上一轮遗留的编译断链（36 错误清零）：空接口链收拢、DownloadEnqueuer 包路径、RecyclerView/WindowInsets import、Callable 归位、FAB id 等
- M-09 严重回归修复：findV9MmanItemByViewKey 由 AND 恒空改为 qb.or() 三形态匹配
- M-02 残留 TOCTOU / M-15 空 describe 收尾
- com/awesapp 零引用第三方桩移出源码树
- L 组 20 项保持完好；Bugly ID 经 -P 注入不入库

+++ app/src/main/java/com/m3man/utils/DownloadManager.java
@@ -220,9 +220,9 @@ public class DownloadManager {
                     if (item != null) {
                         item.setStatus(FileDownloadStatus.completed);
                         item.setProgress(100);
-                        // M99：字段已改 long，去掉 int 强转避免 >2GB 文件尺寸截断
-                        item.setSoFarBytes((int) f.length());
-                        item.setTotalFarBytes((int) f.length());
+                        // M99：实体字段已改为 long，直接存真实字节数，避免 >2GB 截断
+                        item.setSoFarBytes(f.length());
+                        item.setTotalFarBytes(f.length());
                         item.setFinishedDownloadDate(new Date());
                         dataManager.updateV9MmanItem(item);
                     }
